chore: useIsWeatherShown refactor
